requirements.txt
aesara==2.9.3
aiohttp==3.9.5
AppKit==0.2.8
array_api_strict==1.1.1
astunparse==1.6.3
attr==0.3.2
beartype==0.18.5
bokeh==3.4.1
brotli==1.1.0
brotlicffi==1.1.0.0
cairocffi==1.7.0
Carbon==1.1.10
conda==4.3.16
ConfigParser==7.0.0
contextlib2==21.6.0
coremltools==7.2
cryptography==42.0.7
ctypes_snappy==1.03
Cython==3.0.10
deeplearning==0.0.1
defcon==0.10.3
defusedxml==0.7.1
diffusers==0.27.2
dill==0.3.8
disco==1.40.4
dl==0.1.0
docutils==0.21.2
etcd==2.0.8
expecttest==0.2.1
fastparquet==2024.2.0
foo==.1
freetype_py==2.4.0
fs==2.4.16
genapi==0.0.8
glyphsLib==6.7.1
gmpy2==2.1.5
HTMLParser==0.0.2
hypothesis==6.102.1
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isal==1.6.1
jnius==1.1.0
kerchunk==0.2.5
keyring==25.2.1
lark==1.1.9
lxml==5.2.2
lz4==4.3.3
lzmaffi==0.3.0
matchpy==0.5.5
mock==5.1.0
monkeytype==23.3.0
moviepy==1.0.3
mtrand==0.1
multiset==3.1.0
munkres==1.1.4
mypy==1.10.0
nolds==0.5.2
nose==1.3.7
numarray==1.5.1
Numeric==24.2
olefile==0.47
onnxscript==0.1.0.dev20240514
optree==0.11.0
pandas==2.2.2
paramiko==3.4.0
pickle5==0.0.12
pooch==1.8.1
protobuf==5.26.1
psutil==5.9.8
py==1.11.0
pyarrow==16.0.0
pycosat==0.6.6
pydot==2.0.0
pydy==0.7.1
pygit2==1.14.1
pyglet==2.0.15
pygraphviz==1.13
PyInstaller==6.6.0
pymc==5.15.0
pymc3==3.11.5
pynvml==11.5.0
pyOpenSSL==24.1.0
PyQt5==5.15.10
PyQt6==6.7.0
pysat==3.2.0
PySide2==5.15.2.1
PySide6==6.7.0
pytest_timeout==2.3.1
pytorch_lightning==2.2.4
pytz==2024.1
pywin32==306
Quartz==0.0.1.dev0
railroad==0.5.0
redis==5.0.4
reportlab==4.2.0
requests_kerberos==0.14.0
Res==0.1.7
safetensors==0.4.3
sage==0.0.0
scikit_learn==1.4.0
scikits.talkbox==0.2.5
selenium==4.20.0
setuptools_scm==8.1.0
shiboken2==5.15.2.1
shiboken6==6.7.0
sip==6.8.3
slack_sdk==3.27.1
Sphinx==7.3.7
symengine==0.11.0
tabulate==0.9.0
theano==1.0.5
thread==2.0.1
threadpoolctl==3.5.0
torch_xla==2.3.0
torcharrow==0.1.0
torchaudio==2.2.0+cu118
torchrec==0.7.0
torchvision==0.17.0+cu118
tornado==6.4
transformers==4.40.2
trove_classifiers==2024.4.10
ufoLib2==0.16.0
uharfbuzz==0.39.1
ujson==5.10.0
unicodedata2==15.1.0
urllib3_secure_extra==0.1.0
xattr==1.1.0
xmlrpclib==1.0.1
xmlrunner==1.7.7
yarl==1.9.4
z3==0.2.0
zopfli==0.2.3
zstandard==0.22.0
